avcodec/sbcdec: Fix integer overflows in sbc_synthesize_eight()
Fixes: signed integer overflow: 518484152 + 1868182638 cannot be represented in type 'int'
Fixes: 17732/clusterfuzz-testcase-minimized-ffmpeg_AV_CODEC_ID_SBC_fuzzer-5663738132168704

Found-by: continuous fuzzing process https://github.com/google/oss-fuzz/tree/master/projects/ffmpeg

+++ libavcodec/sbcdec.c
@@ -270,14 +270,14 @@ static inline void sbc_synthesize_eight(struct sbc_decoder_state *state,
 
         /* Distribute the new matrix value to the shifted position */
         v[offset[i]] =
-            ( ff_synmatrix8[i][0] * frame->sb_sample[blk][ch][0] +
-              ff_synmatrix8[i][1] * frame->sb_sample[blk][ch][1] +
-              ff_synmatrix8[i][2] * frame->sb_sample[blk][ch][2] +
-              ff_synmatrix8[i][3] * frame->sb_sample[blk][ch][3] +
-              ff_synmatrix8[i][4] * frame->sb_sample[blk][ch][4] +
-              ff_synmatrix8[i][5] * frame->sb_sample[blk][ch][5] +
-              ff_synmatrix8[i][6] * frame->sb_sample[blk][ch][6] +
-              ff_synmatrix8[i][7] * frame->sb_sample[blk][ch][7] ) >> 15;
+             (int)( (unsigned)ff_synmatrix8[i][0] * frame->sb_sample[blk][ch][0] +
+                    (unsigned)ff_synmatrix8[i][1] * frame->sb_sample[blk][ch][1] +
+                    (unsigned)ff_synmatrix8[i][2] * frame->sb_sample[blk][ch][2] +
+                    (unsigned)ff_synmatrix8[i][3] * frame->sb_sample[blk][ch][3] +
+                    (unsigned)ff_synmatrix8[i][4] * frame->sb_sample[blk][ch][4] +
+                    (unsigned)ff_synmatrix8[i][5] * frame->sb_sample[blk][ch][5] +
+                    (unsigned)ff_synmatrix8[i][6] * frame->sb_sample[blk][ch][6] +
+                    (unsigned)ff_synmatrix8[i][7] * frame->sb_sample[blk][ch][7] ) >> 15;
     }
 
     /* Compute the samples */
@@ -286,16 +286,16 @@ static inline void sbc_synthesize_eight(struct sbc_decoder_state *state,
 
         /* Store in output, Q0 */
         AV_WN16A(&output_frame->data[ch][blk * 16 + i * 2], av_clip_int16(
-            ( v[offset[i] + 0] * ff_sbc_proto_8_80m0[idx + 0] +
-              v[offset[k] + 1] * ff_sbc_proto_8_80m1[idx + 0] +
-              v[offset[i] + 2] * ff_sbc_proto_8_80m0[idx + 1] +
-              v[offset[k] + 3] * ff_sbc_proto_8_80m1[idx + 1] +
-              v[offset[i] + 4] * ff_sbc_proto_8_80m0[idx + 2] +
-              v[offset[k] + 5] * ff_sbc_proto_8_80m1[idx + 2] +
-              v[offset[i] + 6] * ff_sbc_proto_8_80m0[idx + 3] +
-              v[offset[k] + 7] * ff_sbc_proto_8_80m1[idx + 3] +
-              v[offset[i] + 8] * ff_sbc_proto_8_80m0[idx + 4] +
-              v[offset[k] + 9] * ff_sbc_proto_8_80m1[idx + 4] ) >> 15));
+         (int)( (unsigned)v[offset[i] + 0] * ff_sbc_proto_8_80m0[idx + 0] +
+                (unsigned)v[offset[k] + 1] * ff_sbc_proto_8_80m1[idx + 0] +
+                (unsigned)v[offset[i] + 2] * ff_sbc_proto_8_80m0[idx + 1] +
+                (unsigned)v[offset[k] + 3] * ff_sbc_proto_8_80m1[idx + 1] +
+                (unsigned)v[offset[i] + 4] * ff_sbc_proto_8_80m0[idx + 2] +
+                (unsigned)v[offset[k] + 5] * ff_sbc_proto_8_80m1[idx + 2] +
+                (unsigned)v[offset[i] + 6] * ff_sbc_proto_8_80m0[idx + 3] +
+                (unsigned)v[offset[k] + 7] * ff_sbc_proto_8_80m1[idx + 3] +
+                (unsigned)v[offset[i] + 8] * ff_sbc_proto_8_80m0[idx + 4] +
+                (unsigned)v[offset[k] + 9] * ff_sbc_proto_8_80m1[idx + 4] ) >> 15));
     }
 }
